requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
hypothesis==6.88.1
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
requests==2.31.0
google-generativeai==0.8.3
